+++ dd-java-agent/instrumentation/jdbc/src/main/java/datadog/trace/instrumentation/jdbc/PreparedStatementInstrumentation.java
@@ -6,10 +6,10 @@ import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.nameSta
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
-import static datadog.trace.instrumentation.jdbc.JDBCDecorator.DATABASE_QUERY;
-import static datadog.trace.instrumentation.jdbc.JDBCDecorator.DECORATE;
-import static datadog.trace.instrumentation.jdbc.JDBCUtils.connectionFromStatement;
-import static datadog.trace.instrumentation.jdbc.JDBCUtils.unwrappedStatement;
+import static datadog.trace.bootstrap.instrumentation.jdbc.JDBCDecorator.DATABASE_QUERY;
+import static datadog.trace.bootstrap.instrumentation.jdbc.JDBCDecorator.DECORATE;
+import static datadog.trace.bootstrap.instrumentation.jdbc.JDBCUtils.connectionFromStatement;
+import static datadog.trace.bootstrap.instrumentation.jdbc.JDBCUtils.unwrappedStatement;
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.not;
@@ -54,13 +54,6 @@ public final class PreparedStatementInstrumentation extends Instrumenter.Tracing
   public ElementMatcher<TypeDescription> typeMatcher() {
     return implementsInterface(named("java.sql.PreparedStatement"))
         .and(not(named("scalikejdbc.DBConnectionAttributesWiredPreparedStatement")));
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      packageName + ".JDBCUtils", packageName + ".JDBCUtils$1", packageName + ".JDBCDecorator",
-    };
   }
 
   @Override
